Add a bunch of instances.
I'm trying to build support for Clay in another package, and I found the lack of
some of the more foundational instances to be a bit of a hinderance.

Added instances for 'Eq', 'Functor', and 'Foldable' for several of the data
types. This is to allow for a bit more flexibility when trying to build things
on top of this package. You can now compare two `Selector` values for equality,
for example. Should you need such a thing.

Also, replaced the custom `Fix` type with a prebaked one from the `fixplate`
package. Aside from leaning on already written code, there is a bunch of
NiceThings(TM) that come with the fixplate package.

I've included a `updateClasses` function as example.

I would like to utilise some features from `lens` as well. But I'm considering
making that a separate package as some people aren't a fan of pulling down the
entire package.